Distinction between Zone and Area, and changes to main
main now detects whenever a new Zone is added on the interface and adds it to the list of zones in the backend
Previous area class is now called Zone and has the same properties, but I created a new class called Area that will contain the properties of a type of area, like aiport, city, etc.
For now the Area class only has two properties : min and max altitude, I don't really know what other properties an area type could have.
